+++ src/bt_init.cpp
@@ -18,7 +18,7 @@ SEXP bt_order_(SEXP, SEXP, SEXP, SEXP);
 SEXP bt_unique_(SEXP, SEXP, SEXP);
 SEXP bt_duplicated_(SEXP, SEXP, SEXP);
 SEXP bt_count_(SEXP, SEXP, SEXP);
-SEXP bt_group_agg_(SEXP, SEXP, SEXP, SEXP, SEXP);
+SEXP bt_group_agg_(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 SEXP bt_match_mask_(SEXP, SEXP, SEXP, SEXP);
 SEXP bt_group_id_(SEXP, SEXP);
 SEXP bt_join_(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
@@ -38,7 +38,7 @@ static const R_CallMethodDef CallEntries[] = {
   { "bt_unique_",     (DL_FUNC) &bt_unique_,     3 },
   { "bt_duplicated_", (DL_FUNC) &bt_duplicated_, 3 },
   { "bt_count_",      (DL_FUNC) &bt_count_,      3 },
-  { "bt_group_agg_",  (DL_FUNC) &bt_group_agg_,  5 },
+  { "bt_group_agg_",  (DL_FUNC) &bt_group_agg_,  6 },
   { "bt_match_mask_", (DL_FUNC) &bt_match_mask_, 4 },
   { "bt_group_id_",   (DL_FUNC) &bt_group_id_,   2 },
   { "bt_join_",         (DL_FUNC) &bt_join_,         7 },